Refactor OSv compiler to make it extensible
OSv compiler was hard-coded to compile Java apps only.
Now, we are preparing OSv compiler for NodeJS apps,
therefore simple refactoring was needed. Functionality
should be same as before.
