+++ pkg/compilers/osv/osv_vmware.go
@@ -1,17 +1,18 @@
 package osv
 
 import (
-	"github.com/emc-advanced-dev/unik/pkg/types"
 	"github.com/emc-advanced-dev/pkg/errors"
+	"github.com/emc-advanced-dev/unik/pkg/types"
 )
 
-type OsvVmwareCompiler struct {}
-
+type OsvVmwareCompiler struct {
+	OSvCompilerBase
+}
 
 const OSV_VMWARE_MEMORY = 512
 
 func (osvCompiler *OsvVmwareCompiler) CompileRawImage(params types.CompileImageParams) (_ *types.RawImage, err error) {
-	resultFile, err := compileRawImage(params, false)
+	resultFile, err := osvCompiler.CreateImage(params, false)
 	if err != nil {
 		return nil, errors.New("failed to compile raw osv image", err)
 	}
@@ -24,8 +25,8 @@ func (osvCompiler *OsvVmwareCompiler) CompileRawImage(params types.CompileImageP
 			DeviceMappings: []types.DeviceMapping{
 				types.DeviceMapping{MountPoint: "/", DeviceName: "/dev/sda1"},
 			},
-			StorageDriver: types.StorageDriver_IDE,
-			VsphereNetworkType: types.VsphereNetworkType_VMXNET3,
+			StorageDriver:         types.StorageDriver_IDE,
+			VsphereNetworkType:    types.VsphereNetworkType_VMXNET3,
 			DefaultInstanceMemory: OSV_VMWARE_MEMORY,
 		},
 	}, nil
