+++ pkg/daemon/daemon.go
@@ -27,12 +27,12 @@ import (
 	"github.com/emc-advanced-dev/unik/pkg/providers/vsphere"
 	"github.com/emc-advanced-dev/unik/pkg/state"
 	"github.com/emc-advanced-dev/unik/pkg/types"
+	"github.com/emc-advanced-dev/unik/pkg/util"
 	"github.com/go-martini/martini"
 	"github.com/layer-x/layerx-commons/lxmartini"
+	"os/exec"
 	"path/filepath"
 	"runtime"
-	"os/exec"
-	"github.com/emc-advanced-dev/unik/pkg/util"
 	"sort"
 )
 
@@ -124,28 +124,28 @@ func NewUnikDaemon(config config.DaemonConfig) (*UnikDaemon, error) {
 
 	_compilers[compilers.RUMP_GO_AWS] = &rump.RumpGoCompiler{
 		RumCompilerBase: rump.RumCompilerBase{
-			DockerImage:   "compilers-rump-go-xen",
-			CreateImage:   rump.CreateImageAws,
+			DockerImage: "compilers-rump-go-xen",
+			CreateImage: rump.CreateImageAws,
 		},
 	}
 	_compilers[compilers.RUMP_GO_VMWARE] = &rump.RumpGoCompiler{
 		RumCompilerBase: rump.RumCompilerBase{
 
-			DockerImage:   "compilers-rump-go-hw",
-			CreateImage:   rump.CreateImageVmware,
+			DockerImage: "compilers-rump-go-hw",
+			CreateImage: rump.CreateImageVmware,
 		},
 	}
 	_compilers[compilers.RUMP_GO_VIRTUALBOX] = &rump.RumpGoCompiler{
 		RumCompilerBase: rump.RumCompilerBase{
 
-			DockerImage:   "compilers-rump-go-hw",
-			CreateImage:   rump.CreateImageVirtualBox,
+			DockerImage: "compilers-rump-go-hw",
+			CreateImage: rump.CreateImageVirtualBox,
 		},
 	}
 	_compilers[compilers.RUMP_GO_QEMU] = &rump.RumpGoCompiler{
 		RumCompilerBase: rump.RumCompilerBase{
-			DockerImage:   "compilers-rump-go-hw-no-stub",
-			CreateImage:   rump.CreateImageQemu,
+			DockerImage: "compilers-rump-go-hw-no-stub",
+			CreateImage: rump.CreateImageQemu,
 		},
 	}
 
@@ -178,8 +178,8 @@ func NewUnikDaemon(config config.DaemonConfig) (*UnikDaemon, error) {
 	}
 	_compilers[compilers.RUMP_NODEJS_QEMU] = &rump.RumpScriptCompiler{
 		RumCompilerBase: rump.RumCompilerBase{
-			DockerImage:   "compilers-rump-nodejs-hw-no-stub",
-			CreateImage:   rump.CreateImageQemu,
+			DockerImage: "compilers-rump-nodejs-hw-no-stub",
+			CreateImage: rump.CreateImageQemu,
 		},
 		RunScriptArgs: "/bootpart/node-wrapper.js",
 	}
@@ -189,9 +189,21 @@ func NewUnikDaemon(config config.DaemonConfig) (*UnikDaemon, error) {
 	_compilers[compilers.RUMP_PYTHON_VMWARE] = rump.NewRumpPythonCompiler("compilers-rump-python3-hw", rump.CreateImageVmwareAddStub, rump.BootstrapTypeUDP)
 	_compilers[compilers.RUMP_PYTHON_QEMU] = rump.NewRumpPythonCompiler("compilers-rump-python3-hw-no-stub", rump.CreateImageQemu, rump.BootstrapTypeUDP)
 
-	_compilers[compilers.OSV_JAVA_AWS] = &osv.OsvAwsCompiler{}
-	_compilers[compilers.OSV_JAVA_VIRTUALBOX] = &osv.OsvVirtualboxCompiler{}
-	_compilers[compilers.OSV_JAVA_VMAWRE] = &osv.OsvVmwareCompiler{}
+	_compilers[compilers.OSV_JAVA_AWS] = &osv.OsvAwsCompiler{
+		OSvCompilerBase: osv.OSvCompilerBase{
+			CreateImage: osv.CreateImageJava,
+		},
+	}
+	_compilers[compilers.OSV_JAVA_VIRTUALBOX] = &osv.OsvVirtualboxCompiler{
+		OSvCompilerBase: osv.OSvCompilerBase{
+			CreateImage: osv.CreateImageJava,
+		},
+	}
+	_compilers[compilers.OSV_JAVA_VMAWRE] = &osv.OsvVmwareCompiler{
+		OSvCompilerBase: osv.OSvCompilerBase{
+			CreateImage: osv.CreateImageJava,
+		},
+	}
 
 	d := &UnikDaemon{
 		server:    lxmartini.QuietMartini(),
@@ -339,7 +351,7 @@ func (d *UnikDaemon) addEndpoints() {
 				"args":         args,
 				"compiler":     compilerType,
 				"provider":     providerType,
-				"noCleanup":     noCleanup,
+				"noCleanup":    noCleanup,
 			}).Debugf("compiling raw image")
 
 			compileParams := types.CompileImageParams{
@@ -354,7 +366,6 @@ func (d *UnikDaemon) addEndpoints() {
 				return nil, http.StatusInternalServerError, errors.New("failed to compile raw image", err)
 			}
 			logrus.Debugf("raw image compiled and saved to " + rawImage.LocalImagePath)
-
 
 			if !noCleanup {
 				defer os.Remove(rawImage.LocalImagePath)
@@ -821,14 +832,14 @@ func (d *UnikDaemon) addEndpoints() {
 }
 
 func streamOutput(outputFunc func() (string, error), w io.Writer) (err error) {
-	defer func(){
+	defer func() {
 		if err != nil {
 			w.Write([]byte(err.Error()))
 		}
 	}()
 
 	//give instance some time to start logs server
-	if err := util.Retry(30, time.Millisecond * 500, func() error {
+	if err := util.Retry(30, time.Millisecond*500, func() error {
 		_, err := outputFunc()
 		return err
 	}); err != nil {
